+++ src/GA_core/__init__.py
@@ -1,3 +1,5 @@
 from GA_core.ga_classes import ProgressObserver, Genetic_Algorithm
 from GA_core.gene import Gene
 from GA_core.chromosome import Chromosome 
+from GA_core.crossover import Crossover_var1, Crossover_var2
+from GA_core.observer import ProgressObserver
