+++ src/GA_core/ga_classes.py
@@ -1,92 +1,4 @@
-from abc import ABC, abstractmethod
 import GA_core as ga_opt
-
-
-class ProgressObserver:
-    """A class representing a progress observer.
-    """
-    def update(self, 
-               generation: int, 
-               best_score: float,
-               best_solution: ga_opt.Chromosome):
-        """Updates observer about the progress of the GA
-
-        Args:
-            generation (int): the number of the generation
-            best_score (float): the fitness of the solution
-            best_solution (Individual_Solution): Individual_Solution object with the best fitness
-        """
-        print(f"Generation {generation}: Best solution - {best_score}")
-
-
-
-class CrossoverInterface(ABC):
-    """A class representing the crossover interface.
-
-    Args:
-        ABC (class): Abstract class
-    """
-
-    @abstractmethod
-    def crossover(self, 
-                  parents: list) -> list:
-        """Placeholder for the corssover functions.
-
-        Args:
-            parents (list): list of Individual_Solution objects representing parent solutions
-
-        Returns:
-            list: list of Individual_Solution objects representing child solutions.
-        """
-        pass
-
-
-class Crossover_var1(CrossoverInterface):
-
-    def crossover(self, 
-                  parents:list) -> None:
-        pass
-
-
-class Crossover_var2(CrossoverInterface):
-
-    def crossover(self, 
-                  parents:list) -> None:
-        pass
-
-
-class Crossover_Factory():
-    """A class implementing the factory design pattern.
-
-    Raises:
-        ValueError: raises an error if an invalid crossover type is provided.
-
-    Returns:
-        list: list of Individual_Solution objects representing child solutions.
-    """
-
-    @staticmethod
-    def create_crossover(type:str, 
-                         parents: list):
-        """Crosses over parent solutions given the specified type
-
-        Args:
-            type (str): type of the crossover
-            parents (list): list of Individual_Solution objects representing parent solutions
-
-        Raises:
-            ValueError: raises an error if an invalid crossover type is provided.
-
-        Returns:
-            list: list of Individual_Solution objects representing child solutions.
-        """
-
-        if type == '1':
-            return Crossover_var1()
-        elif type == '2':
-            return Crossover_var2()
-        
-        else: raise ValueError("nope...")
 
 class Genetic_Algorithm:
     """A class representing a genetic algorithm.
